+++ pebb.so/include/worker_b2b.h
@@ -88,7 +88,7 @@ class pebbworker_b2b : public pebbworker {
   //! default destructor
   virtual ~pebbworker_b2b();
 
-  int initialize(int iSrc, int iDst, bool h2d, bool d2h, size_t Size);
+  int initialize(uint16_t iSrc, uint16_t iDst, bool h2d, bool d2h, size_t Size);
   //! Set back-to-back block size
   void set_b2b_block_sizes(const size_t val) { b2b_block_size = val; }
